fix(upload-progress): drop stray-row fallback (manual sync leak) + monotonic % + nits
Review fixes: SyncButton card syncs no longer leak a permanent mislabeled
uploading row (only caller-upserted rows get Uploading updates; late events are
no-ops); transcode % only advances; drop redundant cast/duplicate import;
clarify queue done = captions-only success.

+++ client/src-tauri/src/commands/library_sync.rs
@@ -151,7 +151,7 @@ pub async fn library_sync_to_cloud(
     // direct to OSS. Best-effort: any failure → no videoKey → iOS falls back to
     // the YouTube embed for this entry.
     let video_key: Option<String> =
-        upload_video(&app, video_dir, &id, &auth_state.session_token, entry.duration_sec as f64).await;
+        upload_video(&app, video_dir, &id, &auth_state.session_token, entry.duration_sec).await;
 
     // 4. POST
     let body = serde_json::json!({
